More 64-bit fixes

+++ driver/PoolAllocator.c
@@ -20,7 +20,7 @@ PoolAllocator createPoolAllocator(void* b, unsigned bs, unsigned s)
 	};
 
 	//initialize linked list of free pointers
-	uint32_t* ptr = pa.nextFreeBlock;
+	uintptr_t* ptr = pa.nextFreeBlock;
 	unsigned last = s/bs - 1;
 	for(unsigned c = 0; c < last; ++c)
 	{
@@ -55,7 +55,7 @@ void* poolAllocate(PoolAllocator* pa)
 	void* ret = pa->nextFreeBlock;
 
 	//set next free block to the one the current next points to
-	pa->nextFreeBlock = *(uint32_t*)pa->nextFreeBlock;
+	pa->nextFreeBlock = *(uintptr_t*)pa->nextFreeBlock;
 
 	return ret;
 }
@@ -66,7 +66,7 @@ void poolFree(PoolAllocator* pa, void* p)
 	assert(p);
 
 	//set block to be freed to point to the current next free block
-	*(uint32_t*)p = (uint32_t)pa->nextFreeBlock;
+	*(uintptr_t*)p = (uintptr_t)pa->nextFreeBlock;
 
 	//set next free block to the freshly freed block
 	pa->nextFreeBlock = p;
